fix(lint): cargo clippy --fix across workspace — remove ~320 warnings
Auto-fixed via cargo clippy --fix:
- 105x div_ceil reimplementation → use built-in .div_ceil()
- 29x redundant closures → point-free style
- 14x Option.and_then(|x| Some(y)) → .map()
- 13x identity operations (+ 0, * 1 etc.) → simplified
- 9x .clone() on Copy types (KeyPoint, Point3, Pose) → removed
- 8x manual assign ops → +=, -=, *=
- 6x useless .into() on same type → removed
- 3x Iterator::find → .iter().find()
- 3x enclosing Ok + ? → direct ?
- 2x manual slice copy → .copy_from_slice()
- plus many more

Manual fixes:
- fix(runtime/shared_memory): add .truncate(true) on file create (real bug)
- fix(features/flann): _max_checks — half-implemented search budget, mark TODO
- fix(hal/fast): simplify identical if/else-if blocks → || operator
- fix(scientific): remove stale ply-rs planning comment from write_ply doc
- fix(python): remove redundant 'use cv_point_cloud;' (path accessible without it)

+++ imgproc/src/histogram.rs
@@ -32,7 +32,7 @@ pub fn histogram_equalization(image: &GrayImage) -> GrayImage {
     let cdf = compute_cdf(&hist);
 
     let cdf_min = cdf.iter().find(|&&x| x > 0).copied().unwrap_or(0);
-    let total = image.width() * image.height() as u32;
+    let total = image.width() * image.height();
 
     let mut lut = [0u8; 256];
     if total > cdf_min {
